requirements.txt
stability-sdk
PILLOW

